From yesterday and today> Continued testing linux webkit control with slow progress, able to load it but not inside avalonia control, want to try using webkit directly instead of this portable library which is an annoying middle man. Implemented linux process watcher using shell script extensions but don't think that's the best solution...Also setup vs code on ubuntu so I can build and step trace which is good. Today I separated out tile pieces so theres a title, detail and content w/ tempalte selector for content so I can continue to try to tweak linux webview. Also added a clipping border control like WPF which works nice.

+++ MonkeyPaste.Avalonia/App.axaml.cs
@@ -8,6 +8,7 @@ using System;
 using System.IO;
 using WebViewControl;
 using MonkeyPaste.Common.Avalonia;
+using MonkeyPaste.Common;
 
 namespace MonkeyPaste.Avalonia {
     [DoNotNotify]
@@ -21,6 +22,7 @@ namespace MonkeyPaste.Avalonia {
         }
 
         private void AppTrayIcon_Clicked(object sender, EventArgs e) {
+            MpAvMainWindowViewModel.Instance.ShowWindowCommand.Execute(null);
             return;
         }
 
@@ -33,17 +35,20 @@ namespace MonkeyPaste.Avalonia {
                     File.Delete(cefLogPath);
                 }
 
-                WebView.Settings.OsrEnabled = true;
-                WebView.Settings.LogFile = "ceflog.txt";
-                //WebView.Settings.EnableErrorLogOnly = true;
+                if(!OperatingSystem.IsLinux()) {
+                    WebView.Settings.OsrEnabled = true;
+                    WebView.Settings.LogFile = "ceflog.txt";
+                    //WebView.Settings.EnableErrorLogOnly = true;
 
-                
-                string cefCacheDir = Path.Combine(Environment.CurrentDirectory, "cefcache");
-                if (Directory.Exists(cefCacheDir)) {
-                    //Directory.CreateDirectory(cefCacheDir);
-                    Directory.Delete(cefCacheDir);
+                    string cefCacheDir = Path.Combine(Environment.CurrentDirectory, "cefcache");
+                    if (Directory.Exists(cefCacheDir)) {
+                        //Directory.CreateDirectory(cefCacheDir);
+                        Directory.Delete(cefCacheDir);
+                    }
+                    //WebView.Settings.CachePath = cefCacheDir;
                 }
-                //WebView.Settings.CachePath = cefCacheDir;
+
+
 
 
                 if (OperatingSystem.IsLinux()) {
@@ -58,9 +63,12 @@ namespace MonkeyPaste.Avalonia {
                 var bootstrapper = new MpAvBootstrapperViewModel();
                 await bootstrapper.InitAsync();
 
+                MpConsole.WriteLine("Loaded");
+
                 //desktop.MainWindow.Close();
 
                 desktop.MainWindow = new MpAvMainWindow();
+                desktop.MainWindow.Show();
             }
 
             base.OnFrameworkInitializationCompleted();
